refactor: 💡 modified field option group set module
This commit will modify the field option group set module
