--- src/modules/form/entities/field-option-groupset.entity.ts
@@ -1,25 +1,0 @@
-import { Column, Entity, JoinTable, ManyToMany } from 'typeorm';
-
-import { FieldOptionGroup } from '../../maintenance/field-option-group/entities/field-option-group.entity';
-import { EntityCoreProps } from 'src/core/entities/entity-core-props';
-
-@Entity('fieldoptiongroupset', { schema: 'public' })
-export class FieldOptionGroupSet extends EntityCoreProps {
-
-  static plural = 'fieldOptionGroupSets';
-
-  @Column('integer', {
-    nullable: false,
-    primary: true,
-    name: 'fieldoptiongroupsetid',
-  })
-  id: number;
-
-  @ManyToMany(
-    type => FieldOptionGroup,
-    fieldOptionGroup => fieldOptionGroup.fieldOptionGroupSets,
-    { nullable: false },
-  )
-  @JoinTable({ name: 'fieldoptiongroupsetmembers' })
-  fieldOptionGroups: FieldOptionGroup[];
-}
