+++ src/modules/form/form.module.ts
@@ -8,7 +8,7 @@ import { FieldDataTypeService } from '../maintenance/field-data-type/services/fi
 import { FieldOptionService } from '../maintenance/field-option/services/field-option.service';
 import { FieldOptionChildrenService } from './services/field-option-children.service';
 import { FieldOptionGroupService } from '../maintenance/field-option-group/services/field-option-group.service';
-import { FieldOptionGroupSetService } from './services/field-option-groupset.service';
+import { FieldOptionGroupSetService } from '../maintenance/field-option-group-set/services/field-option-group-set.service';
 import { FieldOptionMergeService } from './services/field-option-merge.service';
 import { FieldRelationService } from './services/field-relation.service';
 import { FormService } from './services/form.service';
@@ -22,7 +22,7 @@ import { FieldGroupSetController } from '../maintenance/field-group-set/controll
 import { FieldInputTypeController } from '../maintenance/field-input-type/controllers/field-input-type.controller';
 import { FieldOptionChildrenController } from './controllers/field-option-children.controller';
 import { FieldOptionGroupController } from '../maintenance/field-option-group/controllers/field-option-group.controller';
-import { FieldOptionGroupSetController } from './controllers/field-option-groupset.controller';
+import { FieldOptionGroupSetController } from '../maintenance/field-option-group-set/controllers/field-option-group-set.controller';
 import { FieldOptionMergeController } from './controllers/field-option-merge.controller';
 import { FieldOptionController } from '../maintenance/field-option/controllers/field-option.controller';
 import { FieldRelationController } from './controllers/field-relation.controller';
@@ -37,7 +37,7 @@ import { FieldGroupSet } from '../maintenance/field-group-set/entities/field-gro
 import { FieldInputType } from '../maintenance/field-input-type/entities/field-input-type.entity';
 import { FieldOptionChildren } from './entities/field-option-children.entity';
 import { FieldOptionGroup } from '../maintenance/field-option-group/entities/field-option-group.entity';
-import { FieldOptionGroupSet } from './entities/field-option-groupset.entity';
+import { FieldOptionGroupSet } from '../maintenance/field-option-group-set/entities/field-option-group-set.entity';
 import { FieldOptionMerge } from './entities/field-option-merge.entity';
 import { FieldOption } from '../maintenance/field-option/entities/field-option.entity';
 import { FieldRelation } from './entities/field-relation.entity';
