+++ src/modules/maintenance/field-option-group-set/services/field-option-group-set.service.ts
@@ -1,0 +1,17 @@
+import { Injectable } from '@nestjs/common';
+import { FieldOptionGroupSet } from '../entities/field-option-group-set.entity';
+import { BaseService } from 'src/core/services/base.service';
+import { Repository } from 'typeorm';
+import { InjectRepository } from '@nestjs/typeorm';
+
+@Injectable()
+export class FieldOptionGroupSetService extends BaseService<
+FieldOptionGroupSet
+> {
+    constructor(
+        @InjectRepository(FieldOptionGroupSet)
+        fieldOptionGroupSetRepository: Repository<FieldOptionGroupSet>,
+    ) {
+        super(fieldOptionGroupSetRepository, FieldOptionGroupSet);
+    }
+}
